requirements.txt
streamlit
selenium
Pillow
webdriver-manager